モジュールからクラスへの変更とモジュールを削除しクラスにまとめた
